Refactored interfaces module
