+++ moccasin/interfaces/moccasin_GUI.py
@@ -26,9 +26,7 @@ import requests
 import textwrap
 import sys
 import wx
-import wx.xrc
 import wx.lib.dialogs
-import wx.lib.agw.genericmessagedialog as GMD
 from pkg_resources import get_distribution, DistributionNotFound
 
 from pyparsing import ParseException, ParseResults
@@ -38,15 +36,18 @@ sys.path.append('moccasin')
 sys.path.append('../../')
 sys.path.append('../matlab_parser')
 sys.path.append('../converter')
+
 from converter import *
 from matlab_parser import *
+from printDialog import *
 
-#This custom class allows us to import RTF-formatted text into a RichText widget
-import PyRTFParser
-
+#Lexer and formatter for syntax highlighting and export to RTF
 from pygments import highlight
 from pygments.lexers import get_lexer_by_name
 from pygments.formatters import RtfFormatter
+
+#For import of RTF-formatted text into RichText widget
+import PyRTFParser
 
 # -----------------------------------------------------------------------------
 # Helper functions
@@ -155,114 +156,11 @@ def initializePrintingDefaults(self):
 # -----------------------------------------------------------------------------
 
 _BIOCHAM_URL = 'http://lifeware.inria.fr/biocham/online/rest/export'
-_HELP_URL = "https://github.com/sbmlteam/moccasin/blob/setupFix_branch/docs/quickstart.md"
+_HELP_URL = "https://github.com/sbmlteam/moccasin/blob/master/README.md"
 _LICENSE_URL = "https://www.gnu.org/licenses/lgpl.html"
 _VERSION = "Version:  "+ getPackageVersion()
 _IS_OUTPUT_SAVED = False
 _SAVEAS_ODE = False #Used to save the right file format
-_FONTSIZE = 10
-
-# -----------------------------------------------------------------------------
-# Printout class definition
-# -----------------------------------------------------------------------------
-
-#A printout class that is able to print simple text documents.
-#Does not handle page numbers or titles, and it assumes that no
-#lines are longer than what will fit within the page width
-
-class TextDocPrintout(wx.Printout):
-      
-    def __init__(self, text, title, margins):
-        wx.Printout.__init__(self, title)
-        self.lines = text.split('\n')
-        self.margins = margins
-
-
-    def HasPage(self, page):
-        return page <= self.numPages
-
-
-    def GetPageInfo(self):
-        return (1, self.numPages, 1, self.numPages)
-
-
-    def CalculateScale(self, dc):
-        # Scale the DC such that the printout is roughly the same as
-        # the screen scaling.
-        ppiPrinterX, ppiPrinterY = self.GetPPIPrinter()
-        ppiScreenX, ppiScreenY = self.GetPPIScreen()
-        logScale = float(ppiPrinterX)/float(ppiScreenX)
-
-        # Now adjust if the real page size is reduced (such as when
-        # drawing on a scaled wx.MemoryDC in the Print Preview.)  If
-        # page width == DC width then nothing changes, otherwise we
-        # scale down for the DC.
-        pw, ph = self.GetPageSizePixels()
-        dw, dh = dc.GetSize()
-        scale = logScale * float(dw)/float(pw)
-
-        # Set the DC's scale.
-        dc.SetUserScale(scale, scale)
-
-        # Find the logical units per millimeter (for calculating the
-        # margins)
-        self.logUnitsMM = float(ppiPrinterX)/(logScale*25.4)
-
-
-    def CalculateLayout(self, dc):
-        # Determine the position of the margins and the
-        # page/line height
-        topLeft, bottomRight = self.margins
-        dw, dh = dc.GetSize()
-        self.x1 = topLeft.x * self.logUnitsMM
-        self.y1 = topLeft.y * self.logUnitsMM
-        self.x2 = dc.DeviceToLogicalXRel(dw) - bottomRight.x * self.logUnitsMM 
-        self.y2 = dc.DeviceToLogicalYRel(dh) - bottomRight.y * self.logUnitsMM 
-
-        # use a 1mm buffer around the inside of the box, and a few
-        # pixels between each line
-        self.pageHeight = self.y2 - self.y1 - 2*self.logUnitsMM
-        font = wx.Font(_FONTSIZE, wx.TELETYPE, wx.NORMAL, wx.NORMAL)
-        dc.SetFont(font)
-        self.lineHeight = dc.GetCharHeight() 
-        self.linesPerPage = int(self.pageHeight/self.lineHeight)
-
-
-    def OnPreparePrinting(self):
-        # calculate the number of pages
-        dc = self.GetDC()
-        self.CalculateScale(dc)
-        self.CalculateLayout(dc)
-        self.numPages = len(self.lines) / self.linesPerPage
-        if len(self.lines) % self.linesPerPage != 0:
-            self.numPages += 1
-
-
-    def OnPrintPage(self, page):
-        dc = self.GetDC()
-        self.CalculateScale(dc)
-        self.CalculateLayout(dc)
-
-        # draw a page outline at the margin points
-        dc.SetPen(wx.Pen("black", 0))
-        dc.SetBrush(wx.TRANSPARENT_BRUSH)
-        r = wx.RectPP((self.x1, self.y1),
-                      (self.x2, self.y2))
-        dc.DrawRectangleRect(r)
-        dc.SetClippingRect(r)
-
-        # Draw the text lines for this page
-        line = (page-1) * self.linesPerPage
-        x = self.x1 + self.logUnitsMM
-        y = self.y1 + self.logUnitsMM
-        while line < (page * self.linesPerPage):
-            dc.DrawText(self.lines[line], x, y)
-            y += self.lineHeight
-            line += 1
-            if line >= len(self.lines):
-                break
-        return True
-
 
 # -----------------------------------------------------------------------------
 # Graphical User Interface (GUI) definition
